api/requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
Flask==1.1.2
graphviz==0.14.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
plotly==4.12.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
retrying==1.3.3
scipy==1.4.1
six==1.15.0
Werkzeug==1.0.1
tensorflow==2.2.0
keras==2.4.3
